Change hue to expect an integer

+++ src/operations.cc
@@ -185,18 +185,18 @@ namespace sharp {
           0.0, 0.0, 0.0, 1.0));
   }
 
-  VImage Modulate(VImage image, double const brightness, double const saturation, double const hue) {
+  VImage Modulate(VImage image, double const brightness, double const saturation, int const hue) {
     if (HasAlpha(image)) {
       // Separate alpha channel
       VImage alpha = image[image.bands() - 1];
       return RemoveAlpha(image)
         .colourspace(VIPS_INTERPRETATION_LCH)
-        .linear({brightness, saturation, 1}, {0, 0, hue})
+        .linear({brightness, saturation, 1}, {0, 0, static_cast<double>(hue)})
         .bandjoin(alpha);
     } else {
       return image
         .colourspace(VIPS_INTERPRETATION_LCH)
-        .linear({brightness, saturation, 1}, {0, 0, hue});
+        .linear({brightness, saturation, 1}, {0, 0, static_cast<double>(hue)});
     }
   }
